Make node hostnames configurable.
We plumb this through via the kernel commandline, which we manage via
iPXE. A lowkey goal originally was for boot drives to be trivially
cloneable, but that doesn't work already with our monitoring bits.
That said, boot media creation is still easy and takes under a minute,
so I'm not too broken up about it.